requirements.txt
streamlit
pandas
numpy
matplotlib
pillow
tensorflow